# api/control-plane/swagger.yaml
basePath: /
definitions:
...
  adminapi.whitelistEntryResponse:
    properties:
      added_at:
        type: string
      added_by:
        type: string
      effective:
        type: boolean
      peer_id:
        type: string
    type: object
  adminapi.whitelistPostBody:
    properties:
      peer_id:
        type: string
    type: object
  adminapi.whitelistPostResponse:
    properties:
      effective_after:
        type: string
      peer_id:
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8443
info:
  contact: {}
  description: 控制面 HTTP API：节点能力 JWT 签发、blob 存储位置查询、管理后台 API 与 Prometheus 指标。
  title: MediaWorker Control-Plane API
  version: "1.0"
paths:
...
  /v1/admin/whitelist:
    get:
      description: 列出所有白名单条目及其 L4 能力生效状态
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/adminapi.whitelistEntryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - AdminBearer: []
      summary: L4 白名单列表
      tags:
      - admin-whitelist
    post:
      consumes:
      - application/json
      description: 将节点加入 L4 白名单，下一个 JWT 更新周期（≤1h）生效
      parameters:
      - description: 白名单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminapi.whitelistPostBody'
      produces:
      - application/json
      responses:
        "200":
          description: 已存在（幂等）
          schema:
            $ref: '#/definitions/adminapi.whitelistPostResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/adminapi.whitelistPostResponse'
        "400":
          description: 无效 peer_id
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - AdminBearer: []
      summary: 添加 L4 白名单
      tags:
      - admin-whitelist
  /v1/admin/whitelist/{peer_id}:
    delete:
      description: 从白名单移除节点，下一个 JWT 更新周期（≤1h）生效
      parameters:
      - description: libp2p PeerID
        in: path
        name: peer_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: 节点不在白名单中
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - AdminBearer: []
      summary: 移除 L4 白名单
      tags:
      - admin-whitelist
...
securityDefinitions:
  AdminBearer:
    description: 管理后台 UserToken，由 POST /v1/auth/login 签发
    in: header
    name: Authorization
    type: apiKey
  AlertWebhookToken:
    description: Alertmanager webhook 令牌
    in: header
    name: X-Alert-Token
    type: apiKey
  NodeBearer:
    description: 节点能力 JWT，由 POST /v1/node/jwt 签发
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"